Commit para salvar os testes da API

+++ LabAcademiaBiblioteca/Interfaces/IExercicioService.cs
@@ -2,5 +2,5 @@
 
 public interface IExercicioService
 {
-    Task CM_AlterarCargaExercicioAsync(Exercicio p_Exercicio, int p_NovaCarga);
+    Task CM_AlterarCargaExercicioAsync(Exercicio p_Exercicio, double p_NovaCarga);
 }
